+++ RouteCatalog.cs
@@ -17,6 +17,12 @@ namespace HallownestWayfinder
                 Id = "speedrun_5h",
                 Name = SpeedrunRouteDefinition.Name,
                 Steps = SpeedrunRouteDefinition.Steps
+            },
+            new RoutePlan
+            {
+                Id = "grubs_46",
+                Name = GrubRouteDefinition.Name,
+                Steps = GrubRouteDefinition.Steps
             }
         };
     }